requirements.txt
Flask==2.3.2
Flask-Session==0.5.0
pymongo==4.4.0
python-dotenv==1.0.0
pandas==2.2.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tqdm==4.65.0
langdetect==1.0.9
google-cloud-translate==3.11.1
